requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
bcrypt==4.2.0
Brotli==1.1.0
cached-property==1.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.115.0
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
idna==3.10
inflection==0.5.1
Mako==1.3.5
MarkupSafe==2.1.5
mutagen==1.47.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlalchemy-orm==1.2.10
starlette==0.38.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
yt-dlp==2024.10.7
